removed TITLE_DESC/TITLE_ASC from SearchSort in the Users service in favor or more specific sort options.
When searching for a user within elasticsearch the "title" field is likely tokenized in a such a way that
attempting to sort by that field will produce what seems like invalid results.  For example, searching
for 'Adam Zebra' would be sorted first in both scenarios (asc/desc) since elasticsearch will split the
value and have "Adam" and "Zebra" as separate values when using a standard or english analyzer.

Technically it's working as expected but it's non-obvious.  For clarity we'll have separate sort options
for the different fields.   Of course a user with multiple names in first/last will again produce the
same issue but it's much less likely to occur.  For example. "Chad Michael Murray" or "Edward James Almos"
or "Joseph Gordon Levitt" may still sort in what seems like an invalid order.

I probably spelled those actor names incorrectly, calm down, it's fine.  The point is, sorting isn't
perfect when using full-text search unless you use a custom analyzer for the field use need sorting
on.  If you the sort must be perfect, do that in a custom mapper and have the query use that field for sort.