Beta pre-form builder (#2)
* remove photo and video urls as file_ids is now the master

* merging up master

* add auth0 fields to account

* add get user message and is_staff/roles fields to user

* add get user message and is_staff/roles fields to user (allow alphanum and underscore)

* Update README.md

* add "email" to get user request to allow for fetch by email

* adding messages for managing users

* adding search users request/response

* add sort to user search directly.

* gdbots schema updates (node_ref is now an identifier, instead of message-ref)

* gdbots schema updates - adding consistent_read to get node requests

* new pbjc updates, fix node_ref patterns

* no slugs lookup on batch request, just node_refs

* added import user command and now require php 7.1

* add toNodeRef methods to identifiers

* added import solicit and grant/revoke user roles command/event

* add taggable mixin to all messages currently using "tags".  once taggable to rule them all.

* require user_id for role granting/revoking

* use latest gdbots/schemas for user agent free form city

* adding submissions mark as read messages

* bring in latest gdbots/schemas with hints to context conversion

* add request for getting active accounts which will be used in deployment process to create ncr, es, etc.

* get rid of "index-user" commands.  we'll handle those via existing subscribers/projectors, doesn't need a separate command

* removed TITLE_DESC/TITLE_ASC from SearchSort in the Users service in favor or more specific sort options.

When searching for a user within elasticsearch the "title" field is likely tokenized in a such a way that
attempting to sort by that field will produce what seems like invalid results.  For example, searching
for 'Adam Zebra' would be sorted first in both scenarios (asc/desc) since elasticsearch will split the
value and have "Adam" and "Zebra" as separate values when using a standard or english analyzer.

Technically it's working as expected but it's non-obvious.  For clarity we'll have separate sort options
for the different fields.   Of course a user with multiple names in first/last will again produce the
same issue but it's much less likely to occur.  For example. "Chad Michael Murray" or "Edward James Almos"
or "Joseph Gordon Levitt" may still sort in what seems like an invalid order.

I probably spelled those actor names incorrectly, calm down, it's fine.  The point is, sorting isn't
perfect when using full-text search unless you use a custom analyzer for the field use need sorting
on.  If you the sort must be perfect, do that in a custom mapper and have the query use that field for sort.

* RE #1: Add note to submission

* add "gdbots:pbjx:mixin:indexed:v1" to "eme:solicits:event:note-added-to-submission" and
remove the tracked message mixin as we don't need it on this one.

* Fixed phone pattern to allow number to start with "0"

* Revert "Fixed phone pattern to allow number to start with "0""

This reverts commit daad0112e104d68965bc8416e28191836b133c71.

* recompile schemas with pbjc v0.2.3 which fixes some validation issues, doesn't change the generated code.

* use gdbots/schemas ~1.1 constraint and recompile

* Added command and event for rejecting submissions

* remove gdbots/pbj as gdbots/schemas already requires it and recompile

* recompile schemas with pbjc v0.2.6 which optimizes imports and ensure unique class names